requirements.txt
codetiming
datasets
hydra-core==1.3.2
latex2sympy2
numpy==1.26.4
omegaconf
pybind11
ray==2.42.1
regex
sympy
tensorboard
tensordict==0.8.1
transformers==4.57.1
word2number
wandb
pillow==11.3.0
av==14.3.0
qwen_vl_utils==0.0.11
mathruler==0.1.0
pylatexenc==2.10
numba==0.61.2
torchvision==0.22.1
cloudpickle==3.1.1
click==8.2.1
math_verify